refactor(pages/api): split ai-chat.tsx into _-prefixed sibling modules
Extract CORS, screening, audit, rate-limit, Claude client, and shared types
into _ai-chat-{cors,screening,audit,rate-limit,client,types}.ts. Route file
shrinks from 576 → 256 lines; all security comments (screening policy rationale,
fail-closed note, history sanitization comment, cache_control note) preserved.
All 29 route-handler and payload tests pass unchanged.
